requirements.txt
crewai
langchain-mistralai
firecrawl-py
python-dotenv
requests
Pillow
